ledger_lib

Module transport

Low-level transport implementations for communication with ledger devices and nano apps

Transports are gated by transport_X features, while GenericTransport and GenericDevice provide an abstraction over enabled transports.

Structs

BleDevice

BLE connected ledger device

BleInfo

BLE specific device information

BleTransport

Transport for listing and connecting to BLE connected Ledger devices

GenericTransport

GenericTransport for device communication, abstracts underlying transport types

TcpDevice

TCP based device

TcpInfo

TCP device information

TcpTransport

TCP transport implementation for interacting with Speculos via the TCP APDU socket

UsbDevice

USB HID based device

UsbInfo

Basic USB device information

UsbTransport

USB HID based transport

Enums

GenericDevice

GenericDevice for communication with ledger devices, abstracts underlying transport types

Traits

Transport

Transport trait provides an abstract interface for transport implementations.

Type Aliases

PhantomNonSend

A PhantomData to force a type to be !Send