ledger_lib::transport
Struct UsbTransport

pub struct UsbTransport { /* private fields */ }

USB HID based transport

Safety

Due to hidapi this is not thread safe an only one instance must exist in an application. If you don’t need low-level control see crate::LedgerProvider for a tokio based wrapper.

Implementations

impl UsbTransport

pub fn new() -> Result<Self, Error>

Create a new UsbTransport