ledger_lib
Function launch_app

pub async fn launch_app<T>(
    t: T,
    info: <T as Transport>::Info,
    app_name: &str,
    opts: &LaunchAppOpts,
    timeout: Duration,
) -> Result<<T as Transport>::Device, Error>
where
    T: Transport<Info = LedgerInfo, Filters = Filters> + Send,
    <T as Transport>::Device: Send,

Launch an application by name and return a device handle.

This checks whether an application is running, exits this if it is not the desired application, then launches the specified app by name.

WARNING

Due to the constant re-enumeration of devices when changing app contexts, and the lack of reported serial numbers by ledger devices, this is not incredibly reliable. Use at your own risk.