ledger_lib::transport
Trait Transport

pub trait Transport {
    type Filters: Default + Debug;
    type Info: Debug;
    type Device: Exchange;

    // Required methods
    fn list<'life0, 'async_trait>(
        &'life0 mut self,
        filters: Self::Filters,
    ) -> Pin<Box<dyn Future<Output = Result<Vec<LedgerInfo>, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn connect<'life0, 'async_trait>(
        &'life0 mut self,
        info: Self::Info,
    ) -> Pin<Box<dyn Future<Output = Result<Self::Device, Error>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Transport trait provides an abstract interface for transport implementations

Required Associated Types

type Filters: Default + Debug

Connection filters

type Info: Debug

Device information, used for listing and connecting

type Device: Exchange

Device handle for interacting with the device

Required Methods

fn list<'life0, 'async_trait>( &'life0 mut self, filters: Self::Filters, ) -> Pin<Box<dyn Future<Output = Result<Vec<LedgerInfo>, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

List available devices

fn connect<'life0, 'async_trait>( &'life0 mut self, info: Self::Info, ) -> Pin<Box<dyn Future<Output = Result<Self::Device, Error>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Connect to a device using info from a previous list operation